sepperate SimulationState from server
