+++ simulation_state_test.go
